requirements.txt

pandas
# Dev environment
pip
autopep8

# App
Flask
firebase-admin
python-dotenv
